Add files via upload
During the summer of 2022, I Devised a project in the University of Michigan Department of Physics in which I contributed a python script that could automatically plot lines of best fit in a spectrum graph. For the script, we used python and its libraries (such as Matplotlib) to generate visual data of the spectra being plotted against factors such as energy and row.